+++ src/arcemu-world/Player.cpp
@@ -12964,7 +12964,7 @@ void Player::UpdateKnownCurrencies(uint32 itemId, bool apply)
         if(apply)
 		{
             uint64 oldval = GetUInt64Value( PLAYER_FIELD_KNOWN_CURRENCIES );
-            uint64 newval = oldval | ( (( uint32 )1) << (ctEntry->BitIndex-1) );
+            uint64 newval = oldval | (uint64)( (( uint32 )1) << (ctEntry->BitIndex-1) );
             SetUInt64Value( PLAYER_FIELD_KNOWN_CURRENCIES, newval );
 		}
         else
